Getting the business layer up and running

+++ HexMaster.ShortLink.Core/ShortLinks/Models/ShortLinkListItemDto.cs
@@ -8,6 +8,6 @@ namespace HexMaster.ShortLink.Core.Models
         public string ShortCode { get; set; }
         public string EndpointUrl { get; set; }
         public DateTimeOffset? ExpirationOn { get; set; }
-        public int TotalHits { get; set; }
+        public long TotalHits { get; set; }
     }
 }
